+++ src/SignatureParams.h
@@ -3,11 +3,10 @@
 
 #include <string>
 
-struct SignatureParams
-{
+struct SignatureParams {
     std::string inputPath;
     std::string outputPath;
-    uint64_t blockSize = 1024 * 1024; // bytes
+    uint32_t blockSize = 1024 * 1024; // bytes
 
     // settings
     bool debugMode = false;
